+++ src/Rule/Digit.php
@@ -25,10 +25,10 @@ final class Digit extends CharacterClass
         $translator = Policy::getTranslator();
 
         if ($this->getMax() === null) {
-            return $translator->transChoice(
+            return $translator->trans(
                 'There must be at least one digit.|'.
                 'There must be at least %count% digits.',
-                $this->getMin()
+                ['%count%' => $this->getMin()]
             );
         }
 
@@ -39,18 +39,18 @@ final class Digit extends CharacterClass
         }
 
         if ($this->getMin() === 0) {
-            return $translator->transChoice(
+            return $translator->trans(
                 'There must be at most one digit.|'.
                 'There must be at most %count% digits.',
-                $this->getMax()
+                ['%count%' => $this->getMax()]
             );
         }
 
         if ($this->getMin() === $this->getMax()) {
-            return $translator->transChoice(
+            return $translator->trans(
                 'There must be exactly one digit.|'.
                 'There must be exactly %count% digits.',
-                $this->getMin()
+                ['%count%' => $this->getMin()]
             );
         }