+++ src/Rule/LowerCase.php
@@ -98,10 +98,10 @@ final class LowerCase implements RuleInterface
         $translator = Policy::getTranslator();
 
         if ($this->getMax() === null) {
-            return $translator->transChoice(
+            return $translator->trans(
                 'There must be at least one lower case character.|'.
                 'There must be at least %count% lower case characters.',
-                $this->getMin()
+                ['%count%' => $this->getMin()]
             );
         }
 
@@ -112,18 +112,18 @@ final class LowerCase implements RuleInterface
         }
 
         if ($this->getMin() === 0) {
-            return $translator->transChoice(
+            return $translator->trans(
                 'There must be at most one lower case character.|'.
                 'There must be at most %count% lower case characters.',
-                $this->getMax()
+                ['%count%' => $this->getMax()]
             );
         }
 
         if ($this->getMin() === $this->getMax()) {
-            return $translator->transChoice(
+            return $translator->trans(
                 'There must be exactly one lower case character.|'.
                 'There must be exactly %count% lower case characters.',
-                $this->getMin()
+                ['%count%' => $this->getMin()]
             );
         }